pyproject.toml
[tool.poetry]
name = "useless_bot"
version = "2.1"
description = "self-hosted discord bot"
license = "MIT"
authors = ["MRvillager <mrvillager.dev@gmail.com>"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: aiohttp",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.9",
    "Topic :: Communications :: Chat",
]

[tool.poetry.dependencies]
python = "^3.9"
discord-pretty-help = "*"
aiorwlock = "^1"
aiohttp = { version = ">=3.7.4", extras = ["speedups"] }
uvloop = { version = "*", markers = "sys_platform == 'linux'" }
"py-cord" = { git = "https://github.com/Pycord-Development/pycord.git", extras = ["voice"] }

orjson = { version = "^3", optional = true }
aiofiles = { version = "^0.7", optional = true }

[tool.poetry.extras]
json = ["orjson", "aiofiles"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
